Made the StackLifecycleManager into a single class
