+++ navigation-fragment/src/main/java/com/dmcapps/navigationfragment/v7/StackNavigationManagerFragment.java
@@ -2,12 +2,12 @@ package com.dmcapps.navigationfragment.v7;
 
 import com.dmcapps.navigationfragment.R;
 import com.dmcapps.navigationfragment.common.core.ConfigManager;
+import com.dmcapps.navigationfragment.common.core.StackLifecycleManager;
 import com.dmcapps.navigationfragment.common.core.StackManager;
 import com.dmcapps.navigationfragment.common.core.StateManager;
 import com.dmcapps.navigationfragment.common.interfaces.Config;
 import com.dmcapps.navigationfragment.common.interfaces.Navigation;
 import com.dmcapps.navigationfragment.v7.core.NavigationManagerFragment;
-import com.dmcapps.navigationfragment.v7.core.lifecycle.StackLifecycleManager;
 
 /**
  * This Fragment manages the stack of single navigation on fragments.
